// src/app/models/clientes.ts
export class PDV {
    ruta:       string;
    nombre:     string;
    id:         string;
    idWM:       string;
    direccion:  string;
    contacto:   string;
    telefonoContacto: string;
    horaVisita: string;
    latitud:    number;
    longitud:   number;
    visitado:   boolean;

    constructor( ruta: string, id: string, nombre: string, direccion: string, contacto: string, telContacto: string, hora: string, latitud: number, longitud: number, idWM: string ){
      this.ruta = ruta;
      this.nombre = nombre;
      this.id = id;
      this.direccion = direccion;
      this.contacto = contacto;
      this.telefonoContacto = telContacto;
      this.horaVisita = hora;
      this.latitud = latitud;
      this.longitud = longitud;
      this.visitado = false;
      this.idWM = idWM;
    }
}

// Generated by https://quicktype.io

export interface ClientesBD {
  cod_Zon:                       string;
  cod_Clt:                       string;
  nom_Clt:                       string;
  dir_Clt:                       string;
  nom_Cto:                       string;
  num_Tel:                       string;
  longitud:                      number;
  latitud:                       number;
  altitud:                       number;
  fechA_ACTUALIZACION_UBICACION: Date;
  codigo_WM:                     string;
}